# backend/docs/swagger/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "微信小程序打卡项目后端 API。包含用户认证、打卡管理、问卷系统、考试系统、报表导出等功能。",
        "title": "WeCheckin API",
        "contact": {},
        "version": "2.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/admin/admin/batch_force_offline": {
            "post": {
                "tags": [
                    "在线管理员",
                    "管理端 API"
                ],
                "summary": "批量强制下线",
                "parameters": [
                    {
                        "description": "items: [{idStr,token}, ...]",
                        "name": "items",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Resp"
                        }
                    }
                }
            }
        },
...
        "/admin/enroll_clear": {
            "post": {
                "tags": [
                    "打卡管理",
                    "管理端 API"
                ],
                "summary": "清除打卡全部数据",
                "parameters": [
                    {
                        "type": "string",
                        "description": "项目ID",
                        "name": "id",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Resp"
                        }
                    }
                }
            }
        },
        "/admin/enroll_del": {
            "post": {
                "tags": [
                    "打卡管理",
                    "管理端 API"
                ],
                "summary": "删除打卡项目",
                "parameters": [
                    {
                        "type": "string",
                        "description": "项目ID",
                        "name": "id",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Resp"
                        }
                    }
                }
            }
        },
        "/admin/enroll_detail": {
            "get": {
                "tags": [
                    "打卡管理",
                    "管理端 API"
                ],
                "summary": "获取打卡项目详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "项目ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Resp"
                        }
                    }
                }
            }
        },
        "/admin/enroll_edit": {
            "post": {
                "tags": [
                    "打卡管理",
                    "管理端 API"
                ],
                "summary": "编辑打卡项目",
                "parameters": [
                    {
                        "type": "string",
                        "description": "项目ID",
                        "name": "id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "标题",
                        "name": "title",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "分类ID",
                        "name": "cateId",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "分类名称",
                        "name": "cateName",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "开始时间",
                        "name": "startTime",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "结束时间",
                        "name": "endTime",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "排序",
                        "name": "sort",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "封面图URL",
                        "name": "cover",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "描述",
                        "name": "desc",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Resp"
                        }
                    }
                }
            }
        },
...
        "/admin/enroll_join_data_del": {
            "post": {
                "tags": [
                    "打卡管理",
                    "管理端 API"
                ],
                "summary": "删除打卡导出数据",
                "parameters": [
                    {
                        "type": "string",
                        "description": "项目ID",
                        "name": "enrollId",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Resp"
                        }
                    }
                }
            }
        },
        "/admin/enroll_join_data_export": {
            "get": {
                "tags": [
                    "打卡管理",
                    "管理端 API"
                ],
                "summary": "导出打卡数据",
                "parameters": [
                    {
                        "type": "string",
                        "description": "项目ID",
                        "name": "enrollId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Resp"
                        }
                    }
                }
            }
        },
        "/admin/enroll_join_data_get": {
            "get": {
                "tags": [
                    "打卡管理",
                    "管理端 API"
                ],
                "summary": "获取打卡数据导出链接",
                "parameters": [
                    {
                        "type": "string",
                        "description": "项目ID",
                        "name": "enrollId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Resp"
                        }
                    }
                }
            }
        },
        "/admin/enroll_join_del": {
            "post": {
                "tags": [
                    "打卡管理",
                    "管理端 API"
                ],
                "summary": "删除打卡记录",
                "parameters": [
                    {
                        "type": "string",
                        "description": "记录ID",
                        "name": "id",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Resp"
                        }
                    }
                }
            }
        },
        "/admin/enroll_join_list": {
            "get": {
                "parameters": [
                    {
                        "type": "string",
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "每页数量",
                        "name": "size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Resp"
                        }
                    }
                }
            }
        },
        "/admin/enroll_list": {
            "get": {
                "tags": [
                    "打卡管理",
                    "管理端 API"
                ],
                "summary": "获取打卡项目列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "每页数量",
                        "name": "size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "关键词",
                        "name": "keyword",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Resp"
                        }
                    }
                }
            }
        },
        "/admin/enroll_remove_user": {
            "post": {
                "tags": [
                    "打卡管理",
                    "管理端 API"
                ],
                "summary": "从打卡项目中移除用户（删除用户及所有打卡记录）",
                "parameters": [
                    {
                        "type": "string",
                        "description": "项目ID",
                        "name": "enrollId",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "userId",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Resp"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "response.Resp": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "AdminToken": {
            "description": "管理员 Token，格式: \"Bearer {token}\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "ClientToken": {
            "description": "用户 Token，格式: \"Bearer {token}\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}